only append ":." to $PATH if $PATH is not empty
http/cgi implements a default $PATH if empty, try not to clobber that ability,
despite the fact that i consider it a pitfall for unintented behaviour.

honestly I don't think *we* should be mangling $PATH either but whatever, if
someone wanted something secure they wouldn't be running some dinky cgi wrapper
in the first place.